backend/requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
certifi==2023.5.7
charset-normalizer==3.2.0
dj-database-url==2.0.0
Django==4.2.3
django-cors-headers==4.2.0
django-extensions==3.2.3
django-redis==5.3.0
djangorestframework==3.14.0
Faker==19.1.0
gunicorn==20.1.0
idna==3.4
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==5.5.0
typing_extensions==4.7.1
urllib3==2.0.3
